tests/wpt/tests/html/canvas/element/fill-and-stroke-styles/2d.strokeStyle.colormix.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.strokeStyle.colormix</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.strokeStyle.colormix</h1>
<p class="desc">color-mix works as color input</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("color-mix works as color input");
_addTest(function(canvas, ctx) {

  ctx.strokeStyle = "color-mix(in srgb, red, blue)";
  _assertSame(ctx.strokeStyle, 'color(srgb 0.5 0 0.5)', "ctx.strokeStyle", "'color(srgb 0.5 0 0.5)'");
  ctx.strokeStyle = "color-mix(in srgb, red, color(srgb 1 0 0))";
  _assertSame(ctx.strokeStyle, 'color(srgb 1 0 0)', "ctx.strokeStyle", "'color(srgb 1 0 0)'");

});
</script>
